+++ src/Codeception/Module/Lumen.php
@@ -2,7 +2,8 @@
 namespace Codeception\Module;
 
 use Codeception\Configuration;
-use Codeception\Exception\ModuleConfig;
+use Codeception\Exception\ModuleException;
+use Codeception\Exception\ModuleConfigException;
 use Codeception\Lib\Connector\Lumen as LumenConnector;
 use Codeception\Lib\Framework;
 use Codeception\Lib\Interfaces\ActiveRecord;
@@ -113,7 +114,7 @@ class Lumen extends Framework implements ActiveRecord, PartedModule
      * Before hook.
      *
      * @param \Codeception\TestInterface $test
-     * @throws ModuleConfig
+     * @throws ModuleConfigException
      */
     public function _before(TestInterface $test)
     {
@@ -144,14 +145,14 @@ class Lumen extends Framework implements ActiveRecord, PartedModule
     /**
      * Make sure the Lumen bootstrap file exists.
      *
-     * @throws ModuleConfig
+     * @throws ModuleConfigException
      */
     protected function checkBootstrapFileExists()
     {
         $bootstrapFile = $this->config['bootstrap_file'];
 
         if (!file_exists($bootstrapFile)) {
-            throw new ModuleConfig(
+            throw new ModuleConfigException(
                 $this->module,
                 "Lumen bootstrap file not found in $bootstrapFile.\n"
                 . "Please provide a valid path using the 'bootstrap' config param. "